Large refactoring (see commit comments)
• Changed many uses of NSInteger to NSUInteger
• Change NSDictionary helper to use id for keys instead of restricting
them to NSString
• Change NSArray helper and NSDictionary helper to support many more
number types
• Add support for -copyDeep and -copyDeepMutable to NSArray helper and
NSDictionary helper — these recursively copy contents of the set,
setting an immutable or mutable value of each object

+++ Classes/NSWindowHelper.m
@@ -75,8 +75,8 @@ NSString * const NSWindowAutosaveFrameMovesToActiveDisplay = @"NSWindowAutosaveF
 		NSPoint p = NSMakePoint((rect.origin.x + (rect.size.width / 2)), 
 								(rect.origin.y + (rect.size.height / 2)));
 		
-		NSInteger w = size.size.width;
-		NSInteger h = size.size.height;
+		CGFloat w = size.size.width;
+		CGFloat h = size.size.height;
 		
 		rect = NSMakeRect((p.x - (w / 2)), (p.y - (h / 2)), w, h);
